fix: service move service outside model
