+++ lib/service/career_service.dart
@@ -2,8 +2,8 @@ import 'package:capstone_project/model/api/interceptor_api.dart';
 import 'package:capstone_project/model/career_model.dart';
 import 'package:dio/dio.dart';
 
-import '../api/endpoint.dart';
-import '../detail_career_model.dart';
+import '../model/api/endpoint.dart';
+import '../model/detail_career_model.dart';
 
 class CareerService extends InterceptorApi {
   Future<List<CareerModel>> getAllCareer({required String token}) async {
