+++ lib/view/screen/article/article_detail/article_detail_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:capstone_project/model/comment_model.dart';
-import 'package:capstone_project/model/service/reading_list_service.dart';
+import 'package:capstone_project/service/reading_list_service.dart';
 import 'package:capstone_project/utils/state/finite_state.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
