+++ lib/view/screen/auth/login/login_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:capstone_project/model/login_model.dart';
-import 'package:capstone_project/model/service/auth_service.dart';
+import 'package:capstone_project/service/auth_service.dart';
 import 'package:capstone_project/utils/state/finite_state.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
