+++ lib/view/screen/auth/signup/signup_view_model.dart
@@ -1,4 +1,4 @@
-import 'package:capstone_project/model/service/auth_service.dart';
+import 'package:capstone_project/service/auth_service.dart';
 import 'package:capstone_project/utils/state/finite_state.dart';
 import 'package:flutter/material.dart';
