+++ lib/view/screen/counselor_list/counselor_list_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:capstone_project/model/counselor_list_model.dart';
-import 'package:capstone_project/model/service/counselor_list_service.dart';
+import 'package:capstone_project/service/counselor_list_service.dart';
 import 'package:capstone_project/utils/state/finite_state.dart';
 import 'package:flutter/material.dart';
