+++ lib/view/screen/home/home_view_model.dart
@@ -1,7 +1,7 @@
 import 'package:capstone_project/model/forum_model.dart';
 import 'package:capstone_project/model/mock_model.dart';
-import 'package:capstone_project/model/service/home_service.dart';
-import 'package:capstone_project/model/service/user_service.dart';
+import 'package:capstone_project/service/home_service.dart';
+import 'package:capstone_project/service/user_service.dart';
 import 'package:capstone_project/model/user_model.dart';
 import 'package:capstone_project/utils/state/finite_state.dart';
 import 'package:flutter/material.dart';
