+++ lib/view/screen/profile/edit_profile/edit_profile_view_model.dart
@@ -1,4 +1,4 @@
-import 'package:capstone_project/model/service/user_service.dart';
+import 'package:capstone_project/service/user_service.dart';
 import 'package:file_picker/file_picker.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
