+++ lib/view/screen/saved/saved_view_model.dart
@@ -1,5 +1,5 @@
 import 'package:capstone_project/model/reading_list_model.dart';
-import 'package:capstone_project/model/service/reading_list_service.dart';
+import 'package:capstone_project/service/reading_list_service.dart';
 import 'package:flutter/material.dart';
 import 'package:shared_preferences/shared_preferences.dart';
 import '../../../utils/state/finite_state.dart';
