+++ test/view/screen/saved/saved_view_model_test.dart
@@ -1,5 +1,5 @@
 import 'package:capstone_project/model/reading_list_model.dart';
-import 'package:capstone_project/model/service/reading_list_service.dart';
+import 'package:capstone_project/service/reading_list_service.dart';
 import 'package:capstone_project/view/screen/saved/saved_view_model.dart';
 import 'package:flutter_test/flutter_test.dart';
 import 'package:mockito/annotations.dart';
